ios: extract shared BrowseControlsBar header (#942)
The Library screen and the session builder had byte-identical browse
orchestration — type pills, sort menu, tag-filter sheet, and the
button-revealed search bar — each cloning the same activeQuery/activeSort
writes, the same toggle/cancel/applyQuery funcs, and the same
@State/@FocusState. Extract it into one BrowseControlsBar component both
screens render.

Behaviour-preserving: no snapshot churn. Library renders the bar bare;
the builder wraps it with .cardShadow() to separate it from the queue
tray, matching the redesign. previewSearch seeds the revealed-search
state for snapshots.
